Actualizando y preparando todo para la release
