--- CHANGELOG.md.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 5d797903b6dd43c44bc71c7214f4e1cd
-TextScriptImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
