+++ Editor/Evaluacion.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: afa7d9fe819d08d47b3bdfa4c2457224
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
