+++ Editor/IEvaluacion.cs
@@ -1,7 +1,3 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
 namespace ItIsNotOnlyMe.SistemaDeReputacion
 {
     public interface IEvaluacion
@@ -11,33 +7,5 @@ namespace ItIsNotOnlyMe.SistemaDeReputacion
         public void AgregarPreferencia(IPreferencia preferencia);
 
         public void SacarPreferencia(IPreferencia preferencia);
-    }
-
-    public class Evaluacion : IEvaluacion
-    {
-        private List<IPreferencia> _preferencias;
-
-        public Evaluacion()
-        {
-            _preferencias = new List<IPreferencia>();
-        }
-
-        public void AgregarPreferencia(IPreferencia preferencia)
-        {
-            _preferencias.Add(preferencia);
-        }
-
-        public void SacarPreferencia(IPreferencia preferencia)
-        {
-            _preferencias.Remove(preferencia);
-        }
-
-        public float Evaluar(IAccion accion)
-        {
-            float resultado = 0;
-            foreach (IPreferencia preferencia in _preferencias)
-                resultado += preferencia.Alineacion(accion);
-            return resultado;
-        }
     }
 }
